# Conflicts:
#	plugin-sentry/src/main/java/com/yl/lib/plugin/sentry/PrivacySentryPlugin.kt

+++ plugin-sentry/src/main/java/com/yl/lib/plugin/sentry/transform/PrivacyClassProcessor.kt
@@ -3,6 +3,8 @@ package com.yl.lib.plugin.sentry.transform
 import com.yl.lib.plugin.sentry.extension.PrivacyExtension
 import org.apache.commons.io.IOUtils
 import org.gradle.api.Project
+import org.gradle.api.logging.Logger
+import org.gradle.util.GFileUtils
 import org.objectweb.asm.ClassVisitor
 import org.objectweb.asm.Opcodes
 import java.io.File
@@ -12,6 +14,7 @@ import java.io.InputStream
 import java.util.jar.JarFile
 import java.util.jar.JarOutputStream
 import java.util.zip.ZipEntry
+import kotlin.math.log
 
 /**
  * @author yulun
@@ -22,7 +25,7 @@ class PrivacyClassProcessor {
 
     companion object {
 
-        fun runHook(`is`: InputStream?, project: Project): ByteArray? {
+        fun runHook(`is`: InputStream?, extension: PrivacyExtension,logger: Logger): ByteArray? {
             val classReader = org.objectweb.asm.ClassReader(`is`)
 
             // 入参有两个，ClassWriter.COMPUTE_MAXS 和 COMPUTE_FRAMES
@@ -33,9 +36,7 @@ class PrivacyClassProcessor {
             // 定义类访问者
             val classVisitor: ClassVisitor =
                 SentryTraceClassAdapter(
-                    Opcodes.ASM7, classWriter, project.extensions.findByType(
-                        PrivacyExtension::class.java
-                    )
+                    Opcodes.ASM7, classWriter, extension,logger
                 )
             /**
              * ClassReader.SKIP_DEBUG：表示不遍历调试内容，即跳过源文件，源码调试扩展，局部变量表，局部变量类型表和行号表属性，即以下方法既不会被解析也不会被访问（ClassVisitor.visitSource，MethodVisitor.visitLocalVariable，MethodVisitor.visitLineNumber）。使用此标识后，类文件调试信息会被去除，请警记。
@@ -48,7 +49,7 @@ class PrivacyClassProcessor {
             return classWriter.toByteArray()
         }
 
-        fun runCollect(`is`: InputStream?, project: Project): ByteArray? {
+        fun runCollect(`is`: InputStream?, extension: PrivacyExtension,logger: Logger): ByteArray? {
             val classReader = org.objectweb.asm.ClassReader(`is`)
 
             // 入参有两个，ClassWriter.COMPUTE_MAXS 和 COMPUTE_FRAMES
@@ -59,9 +60,7 @@ class PrivacyClassProcessor {
             // 定义类访问者
             val classVisitor: ClassVisitor =
                 CollectHookMethodClassAdapter(
-                    Opcodes.ASM7, classWriter, project.extensions.findByType(
-                        PrivacyExtension::class.java
-                    )
+                    Opcodes.ASM7, classWriter, extension, logger
                 )
             /**
              * ClassReader.SKIP_DEBUG：表示不遍历调试内容，即跳过源文件，源码调试扩展，局部变量表，局部变量类型表和行号表属性，即以下方法既不会被解析也不会被访问（ClassVisitor.visitSource，MethodVisitor.visitLocalVariable，MethodVisitor.visitLineNumber）。使用此标识后，类文件调试信息会被去除，请警记。
@@ -74,7 +73,9 @@ class PrivacyClassProcessor {
             return classWriter.toByteArray()
         }
 
-        fun processJar(project: Project, file: File, extension: PrivacyExtension,runAsm:(InputStream?,Project)->ByteArray?) {
+        private val JAR_SIGNATURE_EXTENSIONS = setOf("SF", "RSA", "DSA", "EC")
+
+        fun processJar(logger: Logger, file: File, extension: PrivacyExtension,runAsm:(InputStream?)->ByteArray?) {
             if (file == null || !file.exists() || !file.name.endsWith(".jar")) {
                 return
             }
@@ -89,47 +90,57 @@ class PrivacyClassProcessor {
             var jarFile = JarFile(file)
             // 以下是标准模板代码
             var enumeration = jarFile.entries()
+
+            var transformSuccess = true
             while (enumeration.hasMoreElements()) {
                 var jarEntry = enumeration.nextElement()
                 // jar里每个元素的名称，对于java来说，就是类名(文件名)
-                var entryName = jarEntry.getName()
+                var entryName = jarEntry.name
+                if (entryName.startsWith("META-INF/") && entryName.substringAfterLast('.') in JAR_SIGNATURE_EXTENSIONS) {
+//                    GFileUtils.copyFile(file,tmpFile)
+                    transformSuccess = false
+                    break
+                }
                 // 封装成zipEntry why？
                 var zipEntry = ZipEntry(entryName)
                 // 针对jarEntry构建输入流
                 var inputStream = jarFile.getInputStream(jarEntry)
                 if (shouldProcessClass(entryName, extension.blackList)) {
-                    project.logger.info("deal with jar file is: $file.absolutePath entryName is $entryName")
+                    logger.info("deal with jar file is: $file.absolutePath entryName is $entryName")
                     jarOutputStream.putNextEntry(zipEntry)
                     // 使用 ASM 对 class 文件进行操控
-                    jarOutputStream.write(runAsm(inputStream, project))
+                    jarOutputStream.write(runAsm(inputStream))
                 } else {
-                    project.logger.info("undeal with jar file is: $file.absolutePath entryName is $entryName")
+                   logger.info("undeal with jar file is: $file.absolutePath entryName is $entryName")
                     // 如果命中黑名单，不做处理，直接输入
                     jarOutputStream.putNextEntry(zipEntry)
                     jarOutputStream.write(IOUtils.toByteArray(inputStream))
                 }
                 jarOutputStream.closeEntry()
             }
-            jarOutputStream.close()
-            jarFile.close()
+            if (transformSuccess) {
+                jarOutputStream.close()
+                jarFile.close()
 
-            if (file.exists()) {
-                file.delete()
+                if (file.exists()) {
+                    file.delete()
+                }
+                // 要把临时文件重命名成源文件的名称
+                tmpFile.renameTo(file)
             }
-            // 要把临时文件重命名成源文件的名称
-            tmpFile.renameTo(file)
+
         }
 
         fun processDirectory(
-            project: Project,
+            logger: Logger,
             inputDir: File,
             inputFile: File,
             extension: PrivacyExtension,
-            runAsm:(InputStream?,Project)->ByteArray?
+            runAsm:(InputStream?)->ByteArray?
         ) {
             if (shouldProcessClass(inputFile.name, extension.blackList)) {
-                project.logger.info("deal with directory file is:" + inputFile.absolutePath)
-                var codeBytes = runAsm(FileInputStream(inputFile), project)
+                logger.info("deal with directory file is:" + inputFile.absolutePath)
+                var codeBytes = runAsm(FileInputStream(inputFile))
                 // 构建输出流，这里是当前目录的原文件；也可以新建个临时文件，写完后再覆盖
                 var fileOutputStream = FileOutputStream(
                     "${inputFile.parent}${File.separator}${inputFile.name}"
@@ -150,15 +161,18 @@ class PrivacyClassProcessor {
 //                || entryName.contains("$") // kotlin object编译后都是内部类，因此这里要放开
                 || entryName.endsWith("R.class")
                 || entryName.endsWith("BuildConfig.class")
-                || entryName.contains("android/support/")
-                || entryName.contains("android/arch/")
-                || entryName.contains("android/app/")
-                || entryName.contains("android/material")
-                || entryName.contains("androidx")
+//                || entryName.contains("android/support/")
+//                || entryName.contains("android/arch/")
+//                || entryName.contains("android/app/")
+//                || entryName.contains("android/material")
+//                || entryName.contains("androidx")
+                || entryName.endsWith(".SF")
+                || entryName.contains(".DSA")
+                || entryName.contains(".RSA")
+                || entryName.contains(".MF")
+
                 // 过滤掉库本身
-                || entryName.contains("com/yl/lib/sentry/hook")
                 || entryName.contains("com/yl/lib/privacy_annotation")
-                || entryName.contains("com/yl/lib/sentry/base")
             ) {
 //            print("checkClassFile className is $entryName false")
                 return false
